+++ src/pages/policy/output/Analysis.jsx
@@ -5,7 +5,7 @@ import Button from "../../../controls/Button";
 import CodeBlock from "layout/CodeBlock";
 import colors from "../../../style/colors";
 import { getParameterAtInstant } from "../../../api/parameters";
-import { BlogContent } from "../../BlogPage";
+import { MarkdownFormatter } from "../../../redesign/style/MarkdownFormatter";
 import { asyncApiCall, countryApiCall } from "../../../api/call";
 import { getImpactReps } from "./ImpactTypes";
 
@@ -57,49 +57,42 @@ export default function Analysis(props) {
   );
   const policyDetails = `I'm using PolicyEngine, a free, open source tool to compute the impact of public policy. I'm writing up an economic analysis of a hypothetical tax-benefit policy reform. Please write the analysis for me using the details below, in their order. You should:
   
-  * First explain each provision of the reform, noting that it's hypothetical and won't represents policy reforms for ${timePeriod} and ${
-    regionKeyToLabel[region]
-  }. Explain how the parameters are changing from the baseline to the reform values using the given data.
-  * Round large numbers like: ${metadata.currency}3.1 billion, ${
-    metadata.currency
-  }300 million, ${metadata.currency}106,000, ${metadata.currency}1.50 (never ${
-    metadata.currency
-  }1.5).
+  * First explain each provision of the reform, noting that it's hypothetical and won't represents policy reforms for ${timePeriod} and ${regionKeyToLabel[region]
+    }. Explain how the parameters are changing from the baseline to the reform values using the given data.
+  * Round large numbers like: ${metadata.currency}3.1 billion, ${metadata.currency
+    }300 million, ${metadata.currency}106,000, ${metadata.currency}1.50 (never ${metadata.currency
+    }1.5).
   * Round percentages to one decimal place.
   * Avoid normative language like 'requires', 'should', 'must', and use quantitative statements over general adjectives and adverbs. If you don't know what something is, don't make it up.
   * Avoid speculating about the intent of the policy or inferring any motives; only describe the observable effects and impacts of the policy. Refrain from using subjective language or making assumptions about the recipients and their needs.
   * Use the active voice where possible; for example, write phrases where the reform is the subject, such as "the reform [or a description of the reform] reduces poverty by x%".
-  * Use ${
-    metadata.countryId === "uk" ? "British" : "American"
-  } English spelling and grammar.
-  * Cite PolicyEngine ${metadata.countryId.toUpperCase()} v${selectedVersion} and the ${
-    metadata.countryId === "uk"
+  * Use ${metadata.countryId === "uk" ? "British" : "American"
+    } English spelling and grammar.
+  * Cite PolicyEngine ${metadata.countryId.toUpperCase()} v${selectedVersion} and the ${metadata.countryId === "uk"
       ? "PolicyEngine-enhanced 2019 Family Resources Survey"
       : "2021 Current Population Survey March Supplement"
-  } microdata when describing policy impacts.
-  * When describing poverty impacts, note that the poverty measure reported is ${
-    metadata.countryId === "uk"
+    } microdata when describing policy impacts.
+  * When describing poverty impacts, note that the poverty measure reported is ${metadata.countryId === "uk"
       ? "absolute poverty before housing costs"
       : "the Supplemental Poverty Measure"
-  }.
+    }.
   * Don't use headers, but do use Markdown formatting. Use - for bullets, and include a newline after each bullet.
   * Include the following embeds inline, without a header so it flows.
   * Immediately after you describe the changes by decile, include the text: {{decileRelativeImpact}}
   * And after the poverty rate changes, include the text: {{povertyImpact}}
-  ${
-    metadata.countryId === "us"
+  ${metadata.countryId === "us"
       ? "* After the racial breakdown of poverty rate changes, include the text: {{racialPovertyImpact}}"
       : ""
-  }
+    }
   * And after the inequality changes, include the text: {{inequalityImpact}}
   * Make sure to accurately represent the changes observed in the data.
 
   This JSON snippet describes the default parameter values: ${JSON.stringify(
-    relevantParameterBaselineValues,
-  )}\n
+      relevantParameterBaselineValues,
+    )}\n
   This JSON snippet describes the baseline and reform policies being compared: ${JSON.stringify(
-    policy,
-  )}\n`;
+      policy,
+    )}\n`;
   const description = `${policyLabel} has the following impacts from the PolicyEngine microsimulation model: 
 
   This JSON snippet describes the relevant parameters with more details: ${JSON.stringify(
@@ -130,15 +123,14 @@ export default function Analysis(props) {
     impact.poverty_by_gender,
   )}
 
-  ${
-    metadata.countryId === "us" &&
+  ${metadata.countryId === "us" &&
     "This JSON describes the baseline and reform poverty impacts by racial group (briefly describe the relative changes): " +
-      JSON.stringify(impact.poverty_by_race.poverty)
-  }
+    JSON.stringify(impact.poverty_by_race.poverty)
+    }
 
   This JSON describes three inequality metrics in the baseline and reform, the Gini coefficient of income inequality, the share of income held by the top 10% of households and the share held by the top 1% (describe the relative changes): ${JSON.stringify(
-    impact.inequality,
-  )}
+      impact.inequality,
+    )}
   
   `;
 
@@ -319,7 +311,7 @@ export default function Analysis(props) {
           />
         )}
         {hasClickedGenerate && analysis && (
-          <BlogContent markdown={analysis} dict={chartDict} />
+          <MarkdownFormatter markdown={analysis} dict={chartDict} />
         )}
       </div>
       <div
